Add smart window gaps
Smart window gaps only add the external and internal gaps when it aids
in visually separating the top levels (when using the tiling mode).

This is done by only adding the outer gap when more than one toplevel
exists in the output and only adding the inner gap when an adjacent
toplevel exists in the given direction.